projects/sportShop/requirements.txt
amqp==5.1.1
asgiref==3.2.10
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.1
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn
idna==3.3
importlib-metadata==4.11.3
kombu==5.2.4
packaging==21.3
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
